Move network port extra DHCP creation tests
Move the extra DHCP options Create call unit and acceptance tests to the
standard networking port package.
